refactor: Introduced BaseTransport abstract class to share common logic.
